requirements.txt
click==5.1
github3.py==0.9.4
clint==0.5.1
